Move FlowOps into a dedicated "flows" package
