--- .sbt_metadata/big_messaging.json
@@ -1,6 +1,0 @@
-{
-  "id" : "big_messaging",
-  "folder" : "common/big_messaging",
-  "dependencyIds" : [
-  ]
-}
