+++ .sbt_metadata/calm_adapter.json
@@ -4,6 +4,6 @@
   "dependencyIds" : [
     "internal_model",
     "source_model_typesafe",
-    "big_messaging"
+    "flows"
   ]
 }
